Break circular dependency.
